fix edit layout scenario routes
