--- app/layout/scenarios/show/features/targets/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
