--- app/layout/scenarios/show/run/chart/constants.ts
@@ -1,38 +1,0 @@
-export const DATA = [
-  {
-    boundaryLength: 1.43,
-    cost: 1.34,
-    isBlm: false,
-    thumbnail: '/images/avatar.png',
-  },
-  {
-    boundaryLength: 1.17,
-    cost: 1.35,
-    isBlm: false,
-    thumbnail: null,
-  },
-  {
-    boundaryLength: 0.49,
-    cost: 1.43,
-    isBlm: true,
-    thumbnail: '/images/avatar.png',
-  },
-  {
-    boundaryLength: 0.32,
-    cost: 1.82,
-    isBlm: false,
-    thumbnail: null,
-  },
-  {
-    boundaryLength: 0.31,
-    cost: 2.1,
-    isBlm: false,
-    thumbnail: '/images/avatar.png',
-  },
-  {
-    boundaryLength: 0.29,
-    cost: 2.89,
-    isBlm: false,
-    thumbnail: '/images/avatar.png',
-  },
-];
