--- app/layout/scenarios/show/run/chart/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
