--- app/layout/scenarios/show/run/field/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
